ai-agents/requirements.txt
anthropic>=0.18.0
httpx>=0.25.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
chromadb==0.4.22
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.6 